server/requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
asgi-csrf==0.8
asgiref==3.3.1
async-generator==1.10
attrs==19.3.0
beautifulsoup4==4.9.3
binarytree==6.3.0
black==19.10b0
bootstrap4==0.1.0
bs4==0.0.1
certifi==2020.12.5
# cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-default-group==1.2.2
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
#datasette==0.56
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==0.5.0
Django==3.1.6
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-configurations==2.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.2.0
django-icons==2.2.1
django-rest-framework==0.1.0
django-taggit==1.3.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.8.0
flake8==3.8.3
future==0.18.2
graphviz==0.19.1
gunicorn==20.0.4
h11==0.12.0
html5lib==1.1
httpcore==0.13.3
httpx==0.18.1
hupper==1.10.2
idna==2.10
immutables==0.16
itsdangerous==1.1.0
itypes==1.2.0
janus==0.6.1
Jinja2==2.11.3
joblib==0.16.0
livereload==2.6.2
lunr==0.5.8
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mergedeep==1.3.4
mock==4.0.3
nltk==3.5
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.0
Pint==0.17
pipenv==2020.11.15
platformdirs==2.5.2
pluggy==0.13.1
ply==3.11
portalocker==2.0.0
# psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==2.1.0
pyparsing==2.4.7
python-baseconv==1.2.2
python-magic==0.4.18
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
qrcode==6.1
rcssmin==1.1.1
regex==2020.7.14
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rfc6555==0.0.0
rjsmin==1.1.0
selectors2==2.0.2
setuptools-scm==6.4.2
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.1
tomli==1.2.3
tornado==6.0.4
tqdm==4.48.0
# typed-ast==1.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
virtualenv==20.3.1
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.1.0
